Create component Footer and Header, add Footer and Header to App.vue
