+++ src/App.vue
@@ -1,10 +1,13 @@
 <script setup>
-import { RouterLink, RouterView } from 'vue-router'
-
+import { RouterView } from 'vue-router'
+import Footer from './components/Footer.vue'
+import Header from './components/Header.vue'
 </script>
 
 <template>
+  <Header />
   <RouterView />
+  <Footer />
 </template>
 
 <style scoped>
